apps/backend/requirements.txt
Flask==3.0.0
Flask-SQLAlchemy==3.1.1
Flask-Migrate==4.0.5
psycopg2-binary==2.9.9
python-dotenv==1.0.0
flask-cors==6.0.1
google-cloud-storage==3.4.1
python-jose[cryptography]==3.3.0
authlib==1.3.0
pytest==8.4.2
pytest-flask==1.3.0
sqlalchemy-cockroachdb==2.0.3
bjoern==3.2.2; sys_platform == "linux"
flake8==7.1.1
black==25.9.0
isort==7.0.0
PyPDF2==3.0.1
python-magic==0.4.27
requests==2.31.0
